Added Authorization & Authorization on /admin Routes

+++ routes/admin/admin.js
@@ -19,7 +19,7 @@ const upload = multer({ storage });
 // GET "/admin/articles" -> List all articles created by admin (both draft and published)
 
 router.get("/articles", (req, res, next) => {
-  Article.find({}, (err, articles) => {
+  Article.find({ userId: req.session.userID }, (err, articles) => {
     if (err) return next(err);
     res.render("admin/admin", { articles });
   });
@@ -43,7 +43,8 @@ router.post("/articles/article/", upload.single("featureImage"), (req, res, next
   Article.create({
     title: req.body.title,
     description: req.body.description,
-    featureImage: req.file
+    featureImage: req.file,
+    userId: req.session.userID
   }, (err, article) => {
       if (err) return next(err);
       res.redirect("/admin/articles")
@@ -55,33 +56,51 @@ router.post("/articles/article/", upload.single("featureImage"), (req, res, next
 router.get("/articles/:id/edit", (req, res, next) => {
   Article.findById(req.params.id, (err, article) => {
     if (err) return next(err);
-    res.render("admin/edit", { article })
+    if (req.session.userID == article.userId) {
+      res.render("admin/edit", { article })
+    } else {
+      res.redirect("/admin/articles")
+    }
   })
 })
 
 // Updating Article
 
 router.post("/articles/:id/update", upload.single("featureImage"), (req, res, next) => {
-  
-  if (req.file) {
-		const base64 = fs.readFileSync(path.join(__dirname, `../../uploads/${req.file.filename}`), { encoding: "base64" });
-    req.file.base64 = base64;
-    req.body.featureImage = req.file
-    
-  }
-  
-  Article.findByIdAndUpdate(req.params.id, req.body, { new: true , useFindAndModify: false},(err, article) => {
+
+  Article.findById(req.params.id, (err, article) => {
     if (err) return next(err);
-    res.redirect("/admin/articles")
+    if (req.session.userID == article.userId) {
+      if (req.file) {
+        const base64 = fs.readFileSync(path.join(__dirname, `../../uploads/${req.file.filename}`), { encoding: "base64" });
+        req.file.base64 = base64;
+        req.body.featureImage = req.file
+      }
+      
+      Article.findByIdAndUpdate(req.params.id, req.body, { new: true, useFindAndModify: false }, (err, article) => {
+        if (err) return next(err);
+        res.redirect("/admin/articles")
+      })
+    } else {
+      res.redirect("/admin/articles")
+    }
   })
+  
 })
 
 // DELETE "/article/:id/delete" 
 
 router.get("/articles/:id/delete", (req, res, next) => {
-  Article.findByIdAndDelete(req.params.id, (err, article) => {
+  Article.findById(req.params.id, (err, article) => {
     if (err) return next(err);
-    res.redirect("/admin/articles")
+    if (req.session.userID == article.userId) {
+      Article.findByIdAndDelete(req.params.id, (err, article) => {
+        if (err) return next(err);
+        res.redirect("/admin/articles")
+      })
+    } else {
+      res.redirect("/admin/articles")
+    }
   })
 })
